corrigiendo los ejercicios y añadido el ejercicio nº 10

+++ Ejercicios Java Tema 4/E6t4.java
@@ -7,7 +7,7 @@
 public class E6t4 { // Clase principal
   public static void main(String[] args) {
   System.out.println("Por favor introduce la Altura");
-  int altura = Integer.parseInt( System.console().readLine());
+  double altura = Integer.parseInt( System.console().readLine());
   double tiempo = (Math.sqrt(altura * 2 / 9.81));
   System.out.println(tiempo + " Es el tiempo en m/s que tarda en caer un objeto a " + altura + " metros de altura");
